Tidied tree and fixed locks
Added a permission string for 'lock' to be used with the lock/unlock
command.
Modified the tree structure for commands, so that contextual commands
and default commands are in separate directories.
Fixed for upstream change to how lock string values are returned.
Made it so that locking and unlocking are on the base object class,
rather than just on lockable doors and such, and because they're in the
same type of call as the 'use' calls, I renamed lock, unlock, and all
the use calls to have a action_ prefix in front of them.
Made lock/unlock into contextual commands, so they only appear in areas
where something is lockable or unlockable.