Overhaul of permission names begun
Unfortunately, this necessitates not calling any built in commands
anymore, because their permission names are hard coded, and in addition
to being uncategorized, there's permissions for too many things, and
some of them collide with obvious choices for in-game stuff like
'lock', which could be used on doors to determine whether players can
use keys on them.

Also, made some changes to the way the bond between players and
characters is handled, due to changes in the evennia code.  Awesome
changes.