Made player suggested changes for page and friends
* Now, in order to have your player be invisible (and to no longer be
able to use friend commands as a result), all of your currently puppeted
characters must be invisible, not just one.

* Now, if you page a character, only that character's connected sessions
are alerted.  If you page/mail a character, it no longer sends a regular
page to all sessions for that player, now it sends a 'you got a page
mail' alert to all sessions instead.