Cleaned up command invocation, added @pref
* Now any 'in character' commands (as in, character actions, not
actions accessible from a character object) can be invoked in the style
of an english sentance, such as 'look at rug' or 'whisper Hello to
dolly', etc.

* In order to facilitate this, I changed follow #stop and lead #stop
into just 'stop', and I added a 'stop' action hook to stop objects.

* I also got rid of follow #auto and replaced it with @pref which has an
autolead and autofollow preference option.  It's also on the player
instead of the character, so I set up a lock function called
owner_lock(<name>) which checks a lock access on the owner of the
character and returns the result.  That way characters can have
individual locks defered to their players.

* I also changed it so that typing a command, then its parameters
without a space no longer matches anything (with the exceptions of say,
and pose, so that : and " still work.

* Tidied up some help entries.