Created character management commands
* Amalgamated the character creation, IC, OOC, and character deletion
commands into one.

* The new character management command now also handles the 'look' when
you're OOC, and it has been much improved.

* Tweaked the way characters are owned.  Now the player and the
character have to mutually acknowledge the ownership bond in their data
in order for ownership to be considered legitimate.  Before, the player
didn't have a list of its characters as attribute data.  Now that it
does, some commands are a lot faster.

* Some minor improvements to the way puppeting and unpuppeting happens
(The player object now holds the routines that check for security, etc.)

* Created a 'owner()' lock function.  This function checks that the
accessor is the owner, or is a character with the same owner.

* Character deletion was added, which doesn't actually delete the
character object, just renders it disabled.

* at_access_failure now only concerns itself with objects accessing
other objects.  This is because you can't tell what session ID is
involved when a player accesses an object, so any messages it sends go
to all sessions.

* Implemented character slots.  There was already a 'max_characters'
limiting the number of new characters, but now if you're already
overbudget with your characters, then you won't be able to puppet some
of them either.  (Starting with the ones you accessed least recently)

* Tightened the rules for player and character names.

* Added 'rename' to the default locks for objects (Same content as the
edit locks.)